+++ 500px/mod.ts
@@ -1,12 +1,13 @@
 /**
  * Interactions with 500px, the photo sharing platform.
  *
- * Provides classes and types, as well as a CLI interface.
+ * Provides classes, functions and types, as well as a CLI interface.
  *
  * @module
  */
 
 import { main } from "./main.ts";
+
 export * from "./client.ts";
 export * from "./data.ts";
 export * from "./types.ts";
