+++ duolingo/README.md
@@ -1,5 +1,9 @@
 # duolingo ([jsr.io](https://jsr.io/@tugrulates/duolingo))
 
+Interactions with Duolingo, the language learning platform.
+
+Provides classes and types, as well as a CLI interface.
+
 ## CLI
 
 Run `duolingo` after installation, or run `deno run -A @tugrulates/duolingo`
@@ -33,3 +37,59 @@ without installation.
 A client for interacting with the Duolingo API.
 
 Requires the JWT (JSON web token) for the logged-in user.
+
+## Types
+
+### [FeedCard](https://jsr.io/@tugrulates/duolingo/doc/~/FeedCard)
+
+A Duolingo feed card, like a milestone or league promotion.
+
+### [Friend](https://jsr.io/@tugrulates/duolingo/doc/~/Friend)
+
+A user other than the current user on Duolingo.
+
+### [League](https://jsr.io/@tugrulates/duolingo/doc/~/League)
+
+A Duolingo league and its user rankings.
+
+### [LeagueUser](https://jsr.io/@tugrulates/duolingo/doc/~/LeagueUser)
+
+A user in a Duolingo league.
+
+## Functions
+
+### [engageWithCard](https://jsr.io/@tugrulates/duolingo/doc/~/engageWithCard)
+
+Engages with the event, following the user or sending a reaction.
+
+### [followLeagueUsers](https://jsr.io/@tugrulates/duolingo/doc/~/followLeagueUsers)
+
+Follows all the users in the league.
+
+### [getEmoji](https://jsr.io/@tugrulates/duolingo/doc/~/getEmoji)
+
+Returns the display emoji for the card.
+
+### [getLeagueUserEmoji](https://jsr.io/@tugrulates/duolingo/doc/~/getLeagueUserEmoji)
+
+Returns the emoji for the user's reaction.
+
+### [getReaction](https://jsr.io/@tugrulates/duolingo/doc/~/getReaction)
+
+Returns the reaction on the card, or picks an appripriate one.
+
+## Constants
+
+### [LANGUAGES](https://jsr.io/@tugrulates/duolingo/doc/~/LANGUAGES)
+
+Language codes on Duolingo, with their names, and flags.
+
+This only lists target languages, and not source languages.
+
+### [LEAGUES](https://jsr.io/@tugrulates/duolingo/doc/~/LEAGUES)
+
+Duolingo leagues tiers, their names, and emojis.
+
+### [REACTIONS](https://jsr.io/@tugrulates/duolingo/doc/~/REACTIONS)
+
+Reactions to Duolingo feed events and the corresponding emojis.
